+++ src/libLookup/Lookup.cpp
@@ -1666,8 +1666,7 @@ bool Lookup::InitMining()
             txBlockRand = {};
 
             m_mediator.m_node->SetState(Node::POW2_SUBMISSION);
-            POW::GetInstance().EthashConfigureLightClient(
-                (uint64_t)m_mediator.m_dsBlockChain.GetBlockCount());
+            POW::GetInstance().EthashConfigureLightClient(m_mediator.m_dsBlockChain.GetBlockCount());
 
             this_thread::sleep_for(chrono::seconds(NEW_NODE_POW_DELAY));
